python-ai/requirements.txt
# Web框架
Flask==3.0.0
flask-cors==4.0.0

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0

# AI/ML库
torch==2.1.2
transformers==4.36.0
pillow==10.1.0
pytesseract==0.3.10
numpy==1.26.2
scikit-learn==1.3.2

# 数据库
pymysql==1.1.0

# 工具库
python-dotenv==1.0.0
requests==2.31.0

# 音频处理（口语评测专用）
librosa>=0.10.1
soundfile>=0.12.1

# 文本转语音
gtts>=2.4.0

# 测试库
pytest==7.4.3
hypothesis==6.92.1
